+++ custom_components/ge_home/devices/fridge.py
@@ -14,7 +14,8 @@ from gehomesdk import (
     ErdFilterStatus,
     HotWaterStatus,
     FridgeModelInfo,
-    ErdConvertableDrawerMode
+    ErdConvertableDrawerMode,
+    ErdDataType
 )
 
 from .base import ApplianceApi
@@ -110,7 +111,7 @@ class FridgeApi(ApplianceApi):
                 GeErdSensor(self, ErdCode.HOT_WATER_SET_TEMP),
                 GeErdPropertySensor(self, ErdCode.HOT_WATER_STATUS, "status", icon_override="mdi:information-outline"),
                 GeErdPropertySensor(self, ErdCode.HOT_WATER_STATUS, "time_until_ready", icon_override="mdi:timer-outline"),
-                GeErdPropertySensor(self, ErdCode.HOT_WATER_STATUS, "current_temp", device_class_override=DEVICE_CLASS_TEMPERATURE),
+                GeErdPropertySensor(self, ErdCode.HOT_WATER_STATUS, "current_temp", device_class_override=DEVICE_CLASS_TEMPERATURE, data_type_override=ErdDataType.INT),
                 GeErdPropertyBinarySensor(self, ErdCode.HOT_WATER_STATUS, "faulted", device_class_override=DEVICE_CLASS_PROBLEM),
                 GeDispenser(self)
             ])