+++ custom_components/ge_home/devices/oven.py
@@ -1,5 +1,6 @@
 import logging
 from typing import List
+from gehomesdk.erd.erd_data_type import ErdDataType
 
 from homeassistant.const import DEVICE_CLASS_POWER_FACTOR
 from homeassistant.helpers.entity import Entity
@@ -92,7 +93,7 @@ class OvenApi(ApplianceApi):
                     cooktop_entities.append(GeErdPropertyBinarySensor(self, ErdCode.COOKTOP_STATUS, prop+".on"))
                     cooktop_entities.append(GeErdPropertyBinarySensor(self, ErdCode.COOKTOP_STATUS, prop+".synchronized"))                    
                     if not v.on_off_only:
-                        cooktop_entities.append(GeErdPropertySensor(self, ErdCode.COOKTOP_STATUS, prop+".power_pct", icon_override="mdi:fire", device_class_override=DEVICE_CLASS_POWER_FACTOR))
+                        cooktop_entities.append(GeErdPropertySensor(self, ErdCode.COOKTOP_STATUS, prop+".power_pct", icon_override="mdi:fire", device_class_override=DEVICE_CLASS_POWER_FACTOR, data_type_override=ErdDataType.INT))
 
         return base_entities + oven_entities + cooktop_entities