+++ custom_components/ge_home/entities/common/ge_erd_sensor.py
@@ -1,5 +1,6 @@
 import logging
 from typing import Optional
+from gehomesdk.erd.erd_data_type import ErdDataType
 from homeassistant.components.sensor import STATE_CLASS_MEASUREMENT
 
 from homeassistant.const import (
@@ -42,23 +43,31 @@ class GeErdSensor(GeErdEntity, Entity):
         device_class_override: str = None,
         state_class_override: str = None,
         uom_override: str = None,
+        data_type_override: ErdDataType = None
     ):
         super().__init__(api, erd_code, erd_override, icon_override, device_class_override)
         self._uom_override = uom_override
         self._state_class_override = state_class_override
+        self._data_type_override = data_type_override
 
     @property
-    def state(self) -> Optional[str]:
+    def native_value(self):
         try:
             value = self.appliance.get_erd_value(self.erd_code)
+
+            # if it's a numeric data type, return it directly            
+            if self._data_type in [ErdDataType.INT, ErdDataType.FLOAT]:
+                return value
+
+            # otherwise, return a stringified version
+            # TODO: perhaps enhance so that there's a list of variables available
+            #       for the stringify function to consume...
+            return self._stringify(value, temp_units=self._temp_units)
         except KeyError:
             return None
-        # TODO: perhaps enhance so that there's a list of variables available
-        #       for the stringify function to consume...
-        return self._stringify(value, temp_units=self._temp_units)
 
     @property
-    def unit_of_measurement(self) -> Optional[str]:
+    def native_unit_of_measurement(self) -> Optional[str]:
         return self._get_uom()
 
     @property
@@ -66,10 +75,22 @@ class GeErdSensor(GeErdEntity, Entity):
         return self._get_state_class()
 
     @property
+    def _data_type(self) -> ErdDataType:
+        if self._data_type_override is not None:
+            return self._data_type_override
+
+        return self.appliance.get_erd_code_data_type(self.erd_code)
+
+    @property
     def _temp_units(self) -> Optional[str]:
-        if self._measurement_system == ErdMeasurementUnits.METRIC:
-            return TEMP_CELSIUS
-        return TEMP_FAHRENHEIT
+        #based on testing, all API values are in Fahrenheit, so we'll redefine
+        #this property to be the configured temperature unit and set the native
+        #unit differently
+        return self.api.hass.config.units.temperature_unit
+
+        #if self._measurement_system == ErdMeasurementUnits.METRIC:
+        #    return TEMP_CELSIUS
+        #return TEMP_FAHRENHEIT
 
     def _get_uom(self):
         """Select appropriate units"""
@@ -83,7 +104,10 @@ class GeErdSensor(GeErdEntity, Entity):
             in [ErdCodeClass.RAW_TEMPERATURE, ErdCodeClass.NON_ZERO_TEMPERATURE]
             or self.device_class == DEVICE_CLASS_TEMPERATURE
         ):
-            return self._temp_units
+            #NOTE: it appears that the API only sets temperature in Fahrenheit,
+            #so we'll hard code this UOM instead of using the device configured
+            #settings
+            return TEMP_FAHRENHEIT
         if (
             self.erd_code_class == ErdCodeClass.BATTERY
             or self.device_class == DEVICE_CLASS_BATTERY
@@ -94,12 +118,12 @@ class GeErdSensor(GeErdEntity, Entity):
         if self.device_class == DEVICE_CLASS_POWER_FACTOR:
             return "%"
         if self.erd_code_class == ErdCodeClass.FLOW_RATE:
-            if self._measurement_system == ErdMeasurementUnits.METRIC:
-                return "lpm"
+            #if self._measurement_system == ErdMeasurementUnits.METRIC:
+            #    return "lpm"
             return "gpm" 
         if self.erd_code_class == ErdCodeClass.LIQUID_VOLUME:       
-            if self._measurement_system == ErdMeasurementUnits.METRIC:
-                return "l"
+            #if self._measurement_system == ErdMeasurementUnits.METRIC:
+            #    return "l"
             return "g"
         return None